moved combat rounds from targeting set to skill schema
