+++ Redninja.Console/Assets/Data/skills.json
@@ -4,55 +4,82 @@
       "dataId": "attack",
       "name": "skill",
       "time": [ 3, 1, 2 ],
-      "targetSetIds": [ "prototype_001" ],
       "baseDamage": 10,
       "critMultiplier": 20,
       "bonusDamageStat": "ATK",
-      "damageTypes": [ "Physical" ]
+      "damageTypes": [ "Physical" ],
+      "targetingSets": [
+        {
+          "targetingRuleId": "single_enemy",
+          "combatRounds": [
+            {
+              "executionStart": 0.25,
+              "operationProviderName": "Damage"
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "dataId": "double_hit",
+      "name": "Double hit",
+      "time": [ 3, 2, 2 ],
+      "baseDamage": 10,
+      "critMultiplier": 20,
+      "bonusDamageStat": "ATK",
+      "damageTypes": [ "Physical" ],
+      "targetingSets": [
+        {
+          "targetingRuleId": "single_enemy",
+          "combatRounds": [
+            {
+              "executionStart": 0.0,
+              "operationProviderName": "Damage"
+            },
+            {
+              "executionStart": 0.50,
+              "operationProviderName": "Damage"
+            }
+          ]
+        }
+      ]
     },
     {
       "dataId": "heal_self",
       "name": "Ejaculate",
       "time": [ 2, 1, 3 ],
-      "targetSetIds": [ "self_only" ],
+      // we'll work on healing definitions later
       "baseDamage": -30,
       "critMultiplier": 1,
       "bonusDamageStat": "ATK",
-      "damageTypes": [ "Physical" ]
+      "damageTypes": [ "Physical" ],
+      "targetingSets": [
+        {
+          "targetingRuleId": "self_only",
+          "combatRounds": []
+        }
+      ]
     }
   ],
-
-  "targetSets": [
+  "targetingRules": [
+    {
+      "dataId": "single_enemy",
+      "targetType": "Entity",
+      "targetTeam": "Enemy",
+      "targetConditionName": "None"
+    },
     {
       "dataId": "prototype_001",
       "pattern": "(0,0)(0,1)",
       "targetType": "Pattern",
-      "targetConditionName": "None",
-      "combatRounds": [
-        {
-          "executionStart": 0.25,
-          "operationProviderName": "Damage",
-          "pattern": "(0,0)"
-        },
-        {
-          "executionStart": 0.25,
-          "operationProviderName": "Damage",
-          "pattern": "(1,0)"
-        }
-      ]
+      "targetTeam": "Enemy",
+      "targetConditionName": "None"
     },
     {
       "dataId": "self_only",
-      "pattern": "(0,0)",
       "targetType": "Entity",
-      "targetConditionName": "None",
-      "combatRounds": [
-        {
-          "executionStart": 1,
-          "operationProviderName": "Damage",
-          "pattern": "(0,0)"
-        }
-      ]
+      "targetTeam": "Self",
+      "targetConditionName": "None"
     }
   ]
 }
