+++ Redninja/Data/Schema/CombatSkillSchema.cs
@@ -8,12 +8,12 @@ namespace Redninja.Data.Schema
 	internal class CombatSkillSchema : IDataSource
 	{
 		public string DataId { get; set; }
-		
+
 		public string Name { get; set; }
 
 		public List<float> Time { get; set; }
 
-		public List<string> TargetSetIds { get; set; }
+		public List<TargetingSetSchema> TargetingSets { get; set; }
 
 		public int BaseDamage { get; set; }
 
@@ -22,5 +22,20 @@ namespace Redninja.Data.Schema
 		public CombatStats BonusDamageStat { get; set; }
 
 		public List<DamageType> DamageTypes { get; set; }
-	}	
+	}
+
+	[Serializable]
+	internal class TargetingSetSchema
+	{
+		public string TargetingRuleId { get; set; }
+		public List<CombatRoundSchema> CombatRounds { get; set; }
+	}
+
+	[Serializable]
+	internal class CombatRoundSchema
+	{
+		public float ExecutionStart { get; set; }
+		public string OperationProviderName { get; set; }
+		public string Pattern { get; set; }
+	}
 }
