+++ Redninja/Data/Schema/TargetingRuleSchema.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Redninja.Components.Targeting;
+
+namespace Redninja.Data.Schema
+{
+	[Serializable]
+	internal class TargetingRuleSchema : IDataSource
+	{
+		public string DataId { get; set; }
+		public string Pattern { get; set; }
+		public TargetTeam TargetTeam { get; set; }
+		public TargetType TargetType { get; set; }
+		public string TargetConditionName { get; set; }
+	}
+}
