+++ json/json.njsproj
@@ -31,6 +31,12 @@
   <ItemGroup>
     <Folder Include="schemas\" />
   </ItemGroup>
+  <ItemGroup>
+    <Content Include="schemas\characters-schema.json" />
+    <Content Include="schemas\encounters-schema.json" />
+    <Content Include="schemas\global-defs.json" />
+    <Content Include="schemas\skills-schema.json" />
+  </ItemGroup>
   <!-- Do not delete the following Import Project.  While this appears to do nothing it is a marker for setting TypeScript properties before our import that depends on them. -->
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\TypeScript\Microsoft.TypeScript.targets" Condition="False" />
   <Import Project="$(VSToolsPath)\Node.js Tools\Microsoft.NodejsTools.targets" />
